feat: App Eraser + OrcaSlicer + PrusaSlicer + SF Symbols + SuperSlicer (#99)
* feat: App Eraser + OrcaSlicer + PrusaSlicer + SF Symbols + SuperSlicer

Added icons for SF Symbols, App Eraser, PrusaSlicer, and OrcaSlicer. SuperSlicer is listed as an alternate name in the PrusaSlicer mapping as they share the same logo.

* fix: PrusaSlicer Icon

After looking at the main PrusaSlicer project, I realized that their logo uses a circle as the background for the logo, as opposed to a squircle as seen in the MacOS app icon. Changed the icon to use a circle instead of a squircle to reflect this.

* fix: PrusaSlicer SVG

Didn't double check the SVG I created on the last commit and noticed it wasn't straight black. Changed the logo color from the redish-grey to black and reoptized it. Apologies.

---------

Co-authored-by: Matthew Bennett <admin@local.matthewben.net>