feat: Phoenix Slides
Added the svg and mapping file for "Phoenix Slides". The app icon is a bit too detailed to just do an outline so I tried to make the wings and lower body a bit larger than the proportions of the app icon to make it look better when shown in a small font.