Added new icons and suggestions for changes (#180)
* Added new icons for various apps

* Changed mapping for gear

second systemeinstellungen has an invisible character after system, which is returned by aerospace for some reason

* New icon for system settings

a suggestion for an more detailed svg which represents the original system settings symbol

* suggestion for preview icon

pdf was to general for me. this represents the loupe in preview app.

* Fixes background of svg

without this change, svg has a non transparent background

* Add another mapping to cisco